pull out gen/run code into utils
